requirements.txt
async_timeout==4.0.2
discord.py==2.2.3
openai==0.27.7
python-dotenv==1.0.0
Requests==2.30.0
youtube_dl==2021.12.17
yt_dlp==2023.3.4
